Create README.md
Lathashree P V is an Assistant Professor in the Department of IT at CMR University, Bangalore, with two years of experience in her current role. She earned her B.E. in Computer Science and Engineering from the Bangalore Institute of Technology in 2018 and her M.Tech in Computer Networking and Engineering from UVCE, Bangalore in 2020. She is currently pursuing a Ph.D in Machine Learning at CMR University.